# uint256: optimize div-related functions by removing unnecessary computation